RAG_/src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.8.1
aiofiles==23.2.1
langchain==0.3.22
PyMuPDF==1.25.5
langchain-community==0.3.20
motor==3.7.0